权限管理 员工授权码区分 门店列表员工分开

+++ app/Http/Controllers/AlipayOpen/UsersController.php
@@ -11,6 +11,7 @@ namespace App\Http\Controllers\AlipayOpen;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class UsersController extends AlipayOpenController
@@ -18,6 +19,11 @@ class UsersController extends AlipayOpenController
 
     public function users()
     {
+        $auth = Auth::user()->can('users');
+        if (!$auth) {
+            echo '你没有权限操作！';
+            die;
+        }
         $user = User::all();
         if ($user) {
             $user = $user->toArray();
@@ -25,8 +31,41 @@ class UsersController extends AlipayOpenController
         return view('admin.alipayopen.users.users', compact("user"));
     }
 
+    //添加用户
+    public function useradd(Request $request)
+    {
+        $data = $request->all();
+        $rules = [
+            'name' => 'required|max:255',
+            'email' => 'required|email|max:255|unique:users',
+            'password' => 'required|min:6|confirmed',
+            'phone' => 'required|min:11|max:11',
+        ];
+        $messages = [
+            'required' => '密码不能为空',
+            'between' => '密码必须是6~20位之间',
+            'confirmed' => '新密码和确认密码不匹配'
+        ];
+        $validator = Validator::make($data, $rules, $messages);
+        if ($validator->fails()) {
+            return back()->withErrors($validator);  //返回一次性错误
+        }
+        User::create([
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'phone' => $data['phone'],
+            'password' => bcrypt($data['password']),
+        ]);
+        return redirect('/admin/alipayopen/users');
+    }
+
     public function updateu(Request $request)
     {
+        $auth = Auth::user()->can('users');
+        if (!$auth) {
+            echo '你没有权限操作！';
+            die;
+        }
         $user = User::where('id', $request->get('id'))->first();
         if ($user) {
             $user = $user->toArray();
@@ -37,14 +76,26 @@ class UsersController extends AlipayOpenController
     //admin  删除账号
     public function deleteu(Request $request)
     {
+        $auth = Auth::user()->can('users');
+        if (!$auth) {
+            echo '你没有权限操作！';
+            die;
+        }
         $user = User::where('id', $request->get('id'))->first();
-        $user->delete();
+        if ($user->name != "admin") {
+            $user->delete();
+        }
         return json_encode(['status' => 1]);
     }
 
     //admin修改账号信息
     public function updateuSave(Request $request)
     {
+        $auth = Auth::user()->can('users');
+        if (!$auth) {
+            echo '你没有权限操作！';
+            die;
+        }
         $email = $request->get('email');
         $user = User::where('email', $email)->first();
         $password_confirm = $request->input('password_confirm');
